refactor data to support unit tests
