+++ src/components/__tests__/UploadPage.spec.ts
@@ -6,8 +6,8 @@ import * as components from "vuetify/components";
 import * as directives from "vuetify/directives";
 import { useUploadedFilesStore } from "../../stores/uploaded-files-store";
 import UploadPage from "../UploadPage.vue";
-import { ImtdTestHelpers } from "../../services/__mocks__/InterMinistryTransferData";
-import { PdfParseError } from "../../services/InterMinistryTransferData";
+import { ImtfdTestHelpers } from "../../data/__mocks__/imtfData";
+import { PdfParseError } from "../../data/imtfData";
 
 // Parts of the pdfjs library used by PdfService are not supported in
 // node.js (they expect to be run in a web browser).  Mock parts of the
@@ -17,7 +17,7 @@ vi.mock("pdfjs-dist", () => ({
   getDocument: vi.fn(),
 }));
 
-vi.mock("../../services/InterMinistryTransferData");
+vi.mock("../../data/imtfData");
 
 global.ResizeObserver = require("resize-observer-polyfill");
 
@@ -57,15 +57,15 @@ describe("UploadPage", () => {
 
   describe("When a valid PDF file is added with the (hidden) input element", () => {
     it("The file is uploaded, parsed and shown on the page", async () => {
-      const spies = ImtdTestHelpers.getSpies();
+      const spies = ImtfdTestHelpers.getSpies();
       const fileInput: any = component.getByLabelText("File input");
       expect(uploadedFilesStore.uploadedFiles.length).toBe(0);
       await fireEvent.change(fileInput, { target: { files: [mockPdfFile] } });
 
       //the file was parsed (no errors were found)
-      expect(spies.importFromPdfSpy).toHaveBeenCalledTimes(1);
-      expect(spies.getMissingRequiredFieldsSpy).toHaveBeenCalledTimes(1);
-      expect(spies.getMissingOptionalFieldsSpy).toHaveBeenCalledTimes(1);
+      expect(spies.importFromPdf).toHaveBeenCalledTimes(1);
+      expect(spies.getMissingRequiredFields).toHaveBeenCalledTimes(1);
+      expect(spies.getMissingOptionalFields).toHaveBeenCalledTimes(1);
 
       //the file is added to the store
       expect(uploadedFilesStore.uploadedFiles.length).toBe(1);
@@ -91,7 +91,7 @@ describe("UploadPage", () => {
 
   describe("When a valid PDF file is dragged-and-dropped onto the 'dropzone'", () => {
     it("The file is uploaded, parsed and shown on the page", async () => {
-      const spies = ImtdTestHelpers.getSpies();
+      const spies = ImtfdTestHelpers.getSpies();
       const dropZone = component.getByText("Drop PDF files here");
 
       expect(uploadedFilesStore.uploadedFiles.length).toBe(0);
@@ -102,9 +102,9 @@ describe("UploadPage", () => {
       });
 
       //the file was parsed (no errors were found)
-      expect(spies.importFromPdfSpy).toHaveBeenCalledTimes(1);
-      expect(spies.getMissingRequiredFieldsSpy).toHaveBeenCalledTimes(1);
-      expect(spies.getMissingOptionalFieldsSpy).toHaveBeenCalledTimes(1);
+      expect(spies.importFromPdf).toHaveBeenCalledTimes(1);
+      expect(spies.getMissingRequiredFields).toHaveBeenCalledTimes(1);
+      expect(spies.getMissingOptionalFields).toHaveBeenCalledTimes(1);
 
       //the file is added to the store
       expect(uploadedFilesStore.uploadedFiles.length).toBe(1);
@@ -132,8 +132,8 @@ describe("UploadPage", () => {
   describe("When an invalid PDF file is dragged-and-dropped onto the 'dropzone'", () => {
     describe("If it's not a pdf", () => {
       it("The file is uploaded, and error shown on the page", async () => {
-        const spies = ImtdTestHelpers.getSpies();
-        spies.importFromPdfSpy.mockImplementation(() => {
+        const spies = ImtfdTestHelpers.getSpies();
+        spies.importFromPdf.mockImplementation(() => {
           throw new PdfParseError("");
         });
         const dropZone = component.getByText("Drop PDF files here");
@@ -146,7 +146,7 @@ describe("UploadPage", () => {
         });
 
         //the file was parsed (no errors were found)
-        expect(spies.importFromPdfSpy).toHaveBeenCalledTimes(1);
+        expect(spies.importFromPdf).toHaveBeenCalledTimes(1);
 
         //the file is added to the store
         expect(uploadedFilesStore.uploadedFiles.length).toBe(1);
@@ -180,8 +180,8 @@ describe("UploadPage", () => {
     });
     describe("If it doesn't have the required fields", () => {
       it("The file is uploaded, and error shown on the page", async () => {
-        const spies = ImtdTestHelpers.getSpies();
-        spies.getMissingRequiredFieldsSpy.mockReturnValueOnce(["field1"]); //errors
+        const spies = ImtfdTestHelpers.getSpies();
+        spies.getMissingRequiredFields.mockReturnValueOnce(["field1"]); //errors
         const dropZone = component.getByText("Drop PDF files here");
 
         expect(uploadedFilesStore.uploadedFiles.length).toBe(0);
@@ -192,9 +192,9 @@ describe("UploadPage", () => {
         });
 
         //the file was parsed (no errors were found)
-        expect(spies.importFromPdfSpy).toHaveBeenCalledTimes(1);
-        expect(spies.getMissingRequiredFieldsSpy).toHaveBeenCalledTimes(1);
-        expect(spies.getMissingOptionalFieldsSpy).toHaveBeenCalledTimes(1);
+        expect(spies.importFromPdf).toHaveBeenCalledTimes(1);
+        expect(spies.getMissingRequiredFields).toHaveBeenCalledTimes(1);
+        expect(spies.getMissingOptionalFields).toHaveBeenCalledTimes(1);
 
         //the file is added to the store
         expect(uploadedFilesStore.uploadedFiles.length).toBe(1);
@@ -228,8 +228,8 @@ describe("UploadPage", () => {
     });
     describe("If it doesn't have the optional fields", () => {
       it("The file is uploaded, and error shown on the page", async () => {
-        const spies = ImtdTestHelpers.getSpies();
-        spies.getMissingOptionalFieldsSpy.mockReturnValueOnce(["field1"]); //warnings
+        const spies = ImtfdTestHelpers.getSpies();
+        spies.getMissingOptionalFields.mockReturnValueOnce(["field1"]); //warnings
         const dropZone = component.getByText("Drop PDF files here");
 
         expect(uploadedFilesStore.uploadedFiles.length).toBe(0);
@@ -240,9 +240,9 @@ describe("UploadPage", () => {
         });
 
         //the file was parsed (no errors were found)
-        expect(spies.importFromPdfSpy).toHaveBeenCalledTimes(1);
-        expect(spies.getMissingRequiredFieldsSpy).toHaveBeenCalledTimes(1);
-        expect(spies.getMissingOptionalFieldsSpy).toHaveBeenCalledTimes(1);
+        expect(spies.importFromPdf).toHaveBeenCalledTimes(1);
+        expect(spies.getMissingRequiredFields).toHaveBeenCalledTimes(1);
+        expect(spies.getMissingOptionalFields).toHaveBeenCalledTimes(1);
 
         //the file is added to the store
         expect(uploadedFilesStore.uploadedFiles.length).toBe(1);
@@ -278,8 +278,8 @@ describe("UploadPage", () => {
 
   describe("When the 'clear' button is clicked", () => {
     it("Removes all previously-added files", async () => {
-      const spies = ImtdTestHelpers.getSpies();
-      spies.importFromPdfSpy.mockImplementation(() => {
+      const spies = ImtfdTestHelpers.getSpies();
+      spies.importFromPdf.mockImplementation(() => {
         throw new PdfParseError("");
       });
